+++ src/pages/match/match.tsx
@@ -1,82 +1,7 @@
-import Card from '@components/card/match-card/card';
-import type { CardProps } from '@components/card/match-card/types/card';
-import FillTabList from '@components/tab/fill-tab/fill-tab-list';
 import TabList from '@components/tab/tab/tab-list';
-import { useState } from 'react';
-
+import MatchTabPanel from '@pages/match/components/match-tab-pannel';
+import { singleMockData, groupMockData } from '@mocks/matchCardData';
 const Match = () => {
-  const [filter, setFilter] = useState<string>('전체');
-
-  const fillTabItems = ['전체', '대기 중', '완료', '실패'];
-
-  const singleMockData: CardProps[] = [
-    {
-      type: 'single',
-      color: 'inactive',
-      nickname: '메이트',
-      date: '2025-07-15',
-      imgUrl: ['/images/profile-1.png'],
-      chips: ['두산', '열정응원러'],
-      awayTeam: '어웨이',
-      homeTeam: '홈',
-      stadium: '경기장',
-      age: 'NN세',
-      gender: '성별',
-      team: 'doosan',
-      style: '열정응원러',
-      status: '매칭 실패',
-    },
-    {
-      type: 'single',
-      color: 'inactive',
-      nickname: '메이트',
-      date: '2025-07-15',
-      imgUrl: ['/images/profile-1.png'],
-      chips: ['두산', '열정응원러'],
-      awayTeam: '어웨이',
-      homeTeam: '홈',
-      stadium: '경기장',
-      age: 'NN세',
-      gender: '성별',
-      team: 'doosan',
-      style: '열정응원러',
-      status: '승인 대기 중',
-    },
-  ];
-
-  const groupMockData: CardProps[] = [
-    {
-      type: 'group',
-      color: 'inactive',
-      count: 3,
-      nickname: '메이트',
-      date: '2025-07-15',
-      imgUrl: ['/images/profile-1.png', '/images/profile-1.png', '/images/profile-1.png'],
-      chips: ['SSG'],
-      awayTeam: '어웨이',
-      homeTeam: '홈',
-      stadium: '경기장',
-      status: '승인 대기 중',
-    },
-    {
-      type: 'group',
-      color: 'active',
-      count: 3,
-      nickname: '메친여자',
-      date: '2025-07-15',
-      imgUrl: ['/images/profile-1.png', '/images/profile-1.png', '/images/profile-1.png'],
-      chips: ['SSG'],
-      awayTeam: '어웨이',
-      homeTeam: '홈',
-      stadium: '경기장',
-      status: '승인 완료',
-    },
-  ];
-
-  const filterCards = (cards: CardProps[]) => {
-    if (filter === '전체') return cards;
-    return cards.filter((card) => card.status?.includes(filter));
-  };
 
   return (
     <div className="flex-col">
@@ -84,38 +9,8 @@ const Match = () => {
         <TabList
           colorMode="light"
           contentMap={{
-            '1:1': (
-              <>
-                <div className="pt-[2.4rem] pb-[2rem]">
-                  <FillTabList tabs={fillTabItems} onChange={setFilter} />
-                </div>
-                <div className="flex-col gap-[0.8rem]">
-                  {filterCards(singleMockData).map((card) => (
-                    <Card
-                      key={`${card.type}-${card.nickname}-${card.date}-${card.status}`}
-                      {...card}
-                      className="w-full"
-                    />
-                  ))}
-                </div>
-              </>
-            ),
-            그룹: (
-              <>
-                <div className="pt-[2.4rem] pb-[2rem]">
-                  <FillTabList tabs={fillTabItems} onChange={setFilter} />
-                </div>
-                <div className="flex-col gap-[0.8rem]">
-                  {filterCards(groupMockData).map((card) => (
-                    <Card
-                      key={`${card.type}-${card.nickname}-${card.date}-${card.status}`}
-                      {...card}
-                      className="w-full"
-                    />
-                  ))}
-                </div>
-              </>
-            ),
+            '1:1': <MatchTabPanel cards={singleMockData} />,
+            그룹: <MatchTabPanel cards={groupMockData} />,
           }}
         />
       </section>
