+++ src/shared/mocks/matchCardData.ts
@@ -1,0 +1,65 @@
+import type { CardProps } from '@components/card/match-card/types/card';
+
+export const singleMockData: CardProps[] = [
+    {
+      type: 'single',
+      color: 'inactive',
+      nickname: '메이트',
+      date: '2025-07-15',
+      imgUrl: ['/images/profile-1.png'],
+      chips: ['두산', '열정응원러'],
+      awayTeam: '어웨이',
+      homeTeam: '홈',
+      stadium: '경기장',
+      age: '20',
+      gender: '여성',
+      team: 'doosan',
+      style: '열정응원러',
+      status: '매칭 실패',
+    },
+    {
+      type: 'single',
+      color: 'inactive',
+      nickname: '메이트',
+      date: '2025-07-15',
+      imgUrl: ['/images/profile-1.png'],
+      chips: ['두산', '열정응원러'],
+      awayTeam: '어웨이',
+      homeTeam: '홈',
+      stadium: '경기장',
+      age: '20',
+      gender: '여성',
+      team: 'doosan',
+      style: '열정응원러',
+      status: '승인 대기 중',
+    },
+  ];
+
+export const groupMockData: CardProps[] = [
+    {
+      type: 'group',
+      color: 'inactive',
+      count: 3,
+      nickname: '메이트',
+      date: '2025-07-15',
+      imgUrl: ['/images/profile-1.png', '/images/profile-1.png', '/images/profile-1.png'],
+      chips: ['SSG'],
+      awayTeam: '어웨이',
+      homeTeam: '홈',
+      stadium: '경기장',
+      status: '승인 대기 중',
+    },
+    {
+      type: 'group',
+      color: 'active',
+      count: 3,
+      nickname: '메친여자',
+      date: '2025-07-15',
+      imgUrl: ['/images/profile-1.png', '/images/profile-1.png', '/images/profile-1.png'],
+      chips: ['SSG'],
+      awayTeam: '어웨이',
+      homeTeam: '홈',
+      stadium: '경기장',
+      status: '승인 완료',
+    },
+  ];
